middleware changes for filter
