+++ main/src/app/api/get_conversation/route.tsx
@@ -1,20 +1,10 @@
 import { NextRequest, NextResponse } from "next/server";
 import dbConnect from "@/lib/db";
 import Conversation from "@/models/Conversation";
-import middleware from "../../middleware";
+import middleware from "../../../middleware";
 import { Message, UserDocument } from "@/datatypes/dataTypes";
 
 export async function GET(req: NextRequest): Promise<Response> {
-
-  // Rate limit check
-  const success = await middleware(req);
-  if (!success) {
-    return NextResponse.json(
-      { error: "Rate limit exceeded. Please try again after a cooldown." },
-      { status: 429 }
-    );
-  }
-
   // GET request check
   if (req.method !== "GET") {
     return NextResponse.json(
