+++ main/src/middleware.ts
@@ -1,0 +1,40 @@
+import { NextResponse } from 'next/server';
+import { NextRequest } from 'next/server';
+import { getToken } from 'next-auth/jwt';
+import { Ratelimit } from '@upstash/ratelimit';
+import { kv } from '@vercel/kv';
+
+const ratelimit = new Ratelimit({
+    redis: kv,
+    limiter: Ratelimit.slidingWindow(20, '1000 s'),
+});
+
+const secret = process.env.NEXTAUTH_SECRET;
+
+const protectedRoutes = ['/chat'];
+
+export default async function middleware(request: NextRequest) {
+    const { pathname } = request.nextUrl;
+
+    const ip = request.ip ?? "127.0.0.1";
+    const { success } = await ratelimit.limit(ip);
+
+    if (!success) {
+        return new NextResponse('Too Many Requests', { status: 429 });
+    }
+
+    if (protectedRoutes.includes(pathname)) {
+        const token = await getToken({ req: request, secret });
+
+        if (!token) {
+            const loginUrl = new URL('/login', request.url);
+            return NextResponse.redirect(loginUrl);
+        }
+    }
+
+    return NextResponse.next();
+}
+
+export const config = {
+    matcher: ['/chat', '/api/:path*'],
+};
